Update Docstrings for CellSignalStrengthLte
Add comments to the docstrings to indicate optional
values that may not always be reported. Users are
still reporting bugs for missing fields, which we
consider optional. This should clarify which methods
are optional vs mandatory for LTE.

For signal strength, RSRP is mandatory, and TA is
mandatory while in LTE_ACTIVE. RSSNR, RSRQ, and CQI
are all optional.

Bug: 36995487
Test: compilation (docstring-only change)
Change-Id: I05af5d42cbf6836109efe6441e95223e81671534