Fix janky wakeup animation
Previous wake-up animation was totally broken before, as multiple
animations were running with different interpolators and different
durations, creating a really jaring experience for the front-door
of the phone, which users see hundreds of times a day.

Instead, we use a single animator to drive everything, and pass in
both the interpolated and the linear value so the right value can
be used in whichever place.

Test: Wakeup/sleep phone with different numbers of notifications
Fixes: 110980608
Change-Id: If1758404a4c49fcd7dc9fa3c93bb18e42a586632