Fixed auto-brightness first screen update.
When auto-brightness is enabled, it should update the screen as soon
as ALS data is available; however, if that data results in the same
screen brightness as before, this update is optimised away, even
though it should be applied (in case the actual screen brightness
did change, e.g. by a user who turned adaptive brightness off).

Test: Turn adaptive brightness off; set the screen brightness to 0;
      turn adaptive brightness on again (do so several times).
      Before the fix, it would often take a while for the screen
      to brighten (because similar ALS data results in the same
      screen brightness as before, an thus no update is issued).
      After the fix, it should brighten immediately.

Fixes: 111381286

Change-Id: I4b6fa2b0efc77e8d01aa10274e9cc33ca8ad583d