Move InputMethodManagerInternal to server package
InputMethodManagerInternal is an internal delegate mechanism for
InputMethodManagerService to expose some internal privileged
operations to other components in the system server process.  Hence
the current package android.view.inputmethod does not make much
sense. Most likely this package was mistakenly chosen when it was
initially introduced by me [1].

Anyway, this is a mechanical refactoring and there should be no
behavior change.

 [1]: I747c37ff6dd8f233faef43f2b5713a4320e848eb
      40b98667ca586cda6fbb79991d8253091f33b3fa

Fix: 112722706
Test: make -j checkbuild
Test: atest CtsInputMethodTestCases CtsInputMethodServiceHostTestCases
Change-Id: I706654eceaa0212d21e053dd25c9d290da512e49