Stop relying on IMM in IMS for token-guarded IME APIs
This is a follow up CL to previous CLs [1][2][3] that made sure that
APIs that are exposed only to IMEs should live in InputMethodService
instead of InputMethodManager.

Now that we have a dedicated Binder inferface [4] that allows
InputMethodService (IMS) to directly send IPCs to
InputMethodManagerService (IMMS) without relying on
InputMethodManager (IMM), it is natural for the above public APIs in
IMS to stop relying on IMM.

This CL also addresses a small concern that it is no longer obvious
when those APIs become available.  Previously, it was a bit more
obvious that passing null IME token doesn't work so IME developers
could imagine that those APIs were unavailable until attachToken() is
called.

With this CL, InputMethodPrivilegedOperations starts showing warning
messages when called too early, which we hope help IME developers
understand why those APIs do nothing when called too early.

  [1]: I3163f3cbe557c85103ca287bee0874a3b4194032
       7471488ae45a24771af566db4480db352f321c8f
  [2]: If6a786c5774805d041ea9672ef2721e4a38df7fc
       e8aa7554fe55ff121115f2b91f94de9d3b8a4a22
  [3]: I6efd5ca473e33e6faeadb7eea7772b9d2b8ca12b
       e348154accc564915b1165048ce1cd9062c75a3a
  [4]: I2f3ec3c5de546fb3603275a4b64000ed3f863b65
       55519a23a6a8a8bb24b6b50b9144b4146e76aaa3

Bug: 114418674
Test: atest CtsInputMethodTestCases CtsInputMethodServiceHostTestCases
Change-Id: I995c4b922f91b94438c1292392b2c3030598594f